CsvFileViewer separate components
