+++ Executables/Kata.UI.Console/CsvFileViewer/CsvFileViewer.cs
@@ -4,30 +4,28 @@
     using System.Collections.Generic;
     using System.Linq;
     using Services.CsvTableizer;
-    using Services.Extensions;
 
     public class CsvFileViewer
     {
-        private const int DefaultPageLength = 24;
-
+        // FEX use record from .Net 5...
+        // or... extract this to an ArgumentDto
+        private readonly (string FileName, int PageLength) settings;
         private readonly CsvTableizerService csvService = new CsvTableizerService();
         private readonly ConsoleKey[] allowedKeys = new[] { ConsoleKey.A, ConsoleKey.X, ConsoleKey.F, ConsoleKey.L, ConsoleKey.N, ConsoleKey.P };
         private readonly string footer = "[N]ext page, [P]revious page, [F]irst page, [L]ast page, e[X]it";
 
-        // FEX use record from .Net 5...
-        // or... extract this to an ArgumentDto
-        private (string fileName, int pageLength) arguments;
         private PageController pageController;
 
-        public void Execute(string[] args)
-        {
-            ////args = new[] { "bla" };
-            args = new[] { "bla", "14" };
 
+        public CsvFileViewer((string fileName, int pageLength) settings) =>
+            this.settings = settings;
+
+
+        public void Execute()
+        {
             int lineCount;
-            this.arguments = this.ReadArguments(args);
-            var csvLines = ReadCsvFile(this.arguments.fileName);
-            this.pageController = new PageController(csvLines.Count, this.arguments.pageLength -2);
+            var csvLines = ReadCsvFile(this.settings.FileName);
+            this.pageController = new PageController(csvLines.Count, this.settings.PageLength -2);
 
             var key = new ConsoleKeyInfo('F', ConsoleKey.F, false, false, false);
             while (key.Key != ConsoleKey.X)
@@ -50,7 +48,7 @@
 
             void printFootLine()
             {
-                while (lineCount <= this.arguments.pageLength) 
+                while (lineCount <= this.settings.PageLength) 
                     writeLine(string.Empty);
                 writeLine(this.footer);
             }
@@ -72,9 +70,9 @@
             return key;
         }
 
-        private List<string> GetTable(List<string> csvLines, ConsoleKey key)
+        private IEnumerable<string> GetTable(IList<string> csvLines, ConsoleKey key)
         {
-            var length = this.arguments.pageLength - 2;
+            var length = this.settings.PageLength - 2;
 
             return key switch
             {
@@ -88,122 +86,7 @@
 
         private static List<string> ReadCsvFile(string fileName)
         {
-            return GetCsvLines().ToList();
-        }
-
-        // TODO extract this to an ArgumentService when it gets larger
-        private (string fileName, int pageLength) ReadArguments(string[] args)
-        {
-            var file = args.Length >= 1 ? args[0] : "file-name not found";
-            var pageLength = args.Length >= 2
-                ? this.GetPageLength(args[1]) 
-                : DefaultPageLength;
-
-            return (file, pageLength);
-        }
-
-
-        private int GetPageLength(string pageLength)
-        {
-            var parsed = int.TryParse(pageLength, out var result);
-            return parsed ? result : DefaultPageLength;
-        }
-
-        private static IEnumerable<string> GetCsvLines()
-        {
-            var line = 0;
-
-            yield return "Name;Street;City;Age";
-            yield return $"{line++} First row -- Meier;Münchener Weg 1;87654 München;65";
-            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
-            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
-            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
-            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
-            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
-            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
-            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
-            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
-            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
-            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
-            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
-            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
-            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
-            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
-            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
-            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
-            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
-            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
-            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
-            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
-            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
-            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
-            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
-            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
-            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
-            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
-            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
-            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
-            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
-            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
-            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
-            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
-            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
-            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
-            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
-            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
-            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
-            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
-            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
-            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
-            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
-            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
-            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
-            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
-            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
-            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
-            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
-            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
-            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
-            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
-            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
-            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
-            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
-            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
-            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
-            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
-            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
-            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
-            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
-            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
-            yield return $"{line++} Last row -- Meier;Münchener Weg 1;87654 München;65";
-        }
-    }
-
-    public class PageController
-    {
-        public PageController(int rowCount, int rowsOnPage) =>
-            this.MaxPages = rowCount / rowsOnPage + 1;
-
-        public int MaxPages { get; set; }
-
-        public int CurrentPage { get; set; }
-
-
-        public int GetFirstPage() => this.CurrentPage = 0;
-
-        public int GetLastPage() => this.CurrentPage = this.MaxPages;
-
-
-        public int GetPrevPage()
-        {
-            this.CurrentPage--;
-            return this.CurrentPage = this.CurrentPage.LimitTo(0, this.MaxPages);
-        }
-
-        public int GetNextPage()
-        {
-            this.CurrentPage++;
-            return this.CurrentPage = this.CurrentPage.LimitTo(0, this.MaxPages);
+            return FakeCsvLines.GetLines().ToList();
         }
     }
 }
