+++ Executables/Kata.UI.Console/CsvFileViewer/FakeCsvLines.cs
@@ -1,0 +1,76 @@
+﻿namespace Kata.UI.Console.CsvFileViewer
+{
+    using System.Collections.Generic;
+
+    public static class FakeCsvLines
+    {
+        public static IEnumerable<string> GetLines()
+        {
+            var line = 0;
+
+            yield return "Name;Street;City;Age";
+            yield return $"{line++} First row -- Meier;Münchener Weg 1;87654 München;65";
+            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
+            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
+            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
+            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
+            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
+            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
+            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
+            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
+            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
+            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
+            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
+            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
+            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
+            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
+            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
+            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
+            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
+            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
+            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
+            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
+            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
+            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
+            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
+            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
+            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
+            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
+            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
+            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
+            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
+            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
+            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
+            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
+            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
+            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
+            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
+            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
+            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
+            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
+            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
+            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
+            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
+            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
+            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
+            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
+            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
+            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
+            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
+            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
+            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
+            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
+            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
+            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
+            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
+            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
+            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
+            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
+            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
+            yield return $"{line++} Peter Pan;Am Hang 5;12345 Einsam;42";
+            yield return $"{line++} Maria Müller;Kölner Straße 45;50123 Köln;43";
+            yield return $"{line++} Paul Meier;Münchener Weg 1;87654 München;65";
+            yield return $"{line++} Last row -- Meier;Münchener Weg 1;87654 München;65";
+        }
+    }
+}
