+++ Executables/Kata.UI.Console/Program.cs
@@ -2,13 +2,35 @@
 {
     public class Program
     {
+
+        private const int DefaultPageLength = 22;
+
+
         static void Main(string[] args)
         {
-            ////args = new[] { "bla" };
+            args = new[] { "bla" };
             ////args = new[] { "bla", "20" };
 
-            var csvFileViewer = new CsvFileViewer.CsvFileViewer();
-            csvFileViewer.Execute(args);
+            var csvFileViewer = new CsvFileViewer.CsvFileViewer(CsvFileViewerSettings(args));
+            csvFileViewer.Execute();
+        }
+
+
+        // TODO extract this to an ArgumentService when it gets larger
+        private static (string fileName, int pageLength) CsvFileViewerSettings(string[] args)
+        {
+            var file = args.Length >= 1 ? args[0] : "file-name not found";
+            var pageLength = args.Length >= 2
+                ? GetPageLength(args[1])
+                : DefaultPageLength;
+
+            return (file, pageLength);
+        }
+
+        private static int GetPageLength(string pageLength)
+        {
+            var parsed = int.TryParse(pageLength, out var result);
+            return parsed ? result : DefaultPageLength;
         }
     }
 }
